added new domain variable, a check to prevent the generation of a new certificate if one already exists and moved away from mageinferno container since nothing is inherited anymore

+++ entry.sh
@@ -1,11 +1,21 @@
 #!/bin/bash
 
 cd /etc/nginx/ssl
-openssl genrsa -des3 -passout pass:x -out nginx.pass.key 2048
-openssl rsa -passin pass:x -in nginx.pass.key -out nginx.key
-rm nginx.pass.key
-openssl req -new -key nginx.key -out nginx.csr -subj "/C=US/ST=California/L=Beverly Hills/O=localhost/OU=localhost/CN=localhost"
-openssl x509 -req -sha256 -days 365 -in nginx.csr -signkey nginx.key -out nginx.crt
+
+if [ ! -f ${APP_DOMAIN}.crt ]; then
+    openssl genrsa -des3 -passout pass:x -out nginx.pass.key 2048
+    openssl rsa -passin pass:x -in nginx.pass.key -out ${APP_DOMAIN}.key
+    rm nginx.pass.key
+    openssl req -new -key ${APP_DOMAIN}.key -out ${APP_DOMAIN}.csr -subj "/C=US/ST=California/L=Beverly Hills/O=${APP_DOMAIN}/OU=${APP_DOMAIN}/CN=${APP_DOMAIN}"
+    openssl x509 -req -sha256 -days 365 -in ${APP_DOMAIN}.csr -signkey ${APP_DOMAIN}.key -out ${APP_DOMAIN}.crt
+fi
+
+for VAR in PHP_HOST PHP_PORT APP_MAGE_MODE APP_DOMAIN
+do
+    VALUE=${!VAR}
+    [ ! -z "${VALUE}" ] && sed -i "s/${VAR}/${VALUE}/" /etc/nginx/conf.d/default.conf
+done
+
 cd /var/www/html
 
-/usr/local/bin/start.sh
+/usr/sbin/nginx -g "daemon off;"
